requirements.txt
keras==2.10.0
numpy==1.23.5
opencv_contrib_python==4.7.0.72
Pillow==9.5.0
tensorflow==2.10.0
tqdm==4.64.1
